Fix: Features wrong folder
